+++ example/pages/sortButton/index.js
@@ -1,0 +1,10 @@
+import Toast from '../../dist/toast/toast'
+
+Page({
+  data: {
+    value: 0
+  },
+  handleClick (event) {
+    Toast(`当前状态：${event.detail}`)
+  }
+})