+++ example/pages/tooltip/index.js
@@ -15,5 +15,15 @@ Page({
   },
   hide () {
     this.setData({ show: !this.data.show })
+  },
+  onShow () {
+    console.log('显示')
+  },
+  onHide () {
+    console.log('隐藏')
+  },
+  link (param) {
+    const data = param.detail.item
+    console.log('调用', data)
   }
 })